RAG/hr-helpdesk/experiments/pyproject.toml
[project]
name = "experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt>=0.9",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-vertexai>=3.2.2",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "streamlit>=1.55.0",
    "tqdm>=4.67.3",
]
